Fix warning.
GitOrigin-RevId: b868b965e8ce8c44540ace16ebde385b3def0076

+++ td/telegram/ContactsManager.cpp
@@ -7496,7 +7496,7 @@ void ContactsManager::on_get_channel_participants_success(
             bot_user_ids.push_back(participant.user_id);
           }
         }
-        administrator_count = administrator_user_ids.size();
+        administrator_count = narrow_cast<int32>(administrator_user_ids.size());
       } else if (filter.is_administrators()) {
         administrator_user_ids = std::move(user_ids);
       } else if (filter.is_bots()) {
